==> scripts/02-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from Open Data Toronto
# Author: Sarah Ding
# Date: 29 November 2024 
# Contact: sarah.ding@mail.utoronto.ca
# License: UofT
# Pre-requisites: installation of package `tidyverse`
# Any other information needed?: N/A


#### Workspace setup ####
library(opendatatoronto)
library(tidyverse)

#### Download data ####

# get package
package <- show_package("ac77f532-f18b-427c-905c-4ae87ce69c93")
package

# get all resources for this package
resources <- list_package_resources("ac77f532-f18b-427c-905c-4ae87ce69c93")

# identify datastore resources; by default, Toronto Open Data sets datastore resource format to CSV for non-geospatial and GeoJSON for geospatial resources
datastore_resources <- filter(resources, tolower(format) %in% c('csv', 'geojson'))

# load the first datastore resource as a sample
data <- filter(datastore_resources, row_number()==1) %>% get_resource()
data

#### Save data ####
write_csv(data, "data/01-raw_data/raw_data.csv")
